requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
requests==2.31.0
requests-mock==1.11.0
python-dotenv==1.0.1
cryptography==42.0.8
PyJWT==2.8.0
openai==1.35.7
websockets==12.0
psutil==5.9.0
click==8.1.7